requirements.txt
streamlit
snowflake-snowpark-python[pandas]
